Add more constructors to accptr_t to have windows happy. Changed nAccelerators to return size_t.

+++ src/core/Process-impl.h
@@ -3,7 +3,7 @@
 
 namespace __impl { namespace core {
 
-inline unsigned
+inline size_t
 Process::nAccelerators() const
 {
     return accs_.size();
